优化代码

+++ Models/LinearClassifier/LogisticRegression.py
@@ -79,7 +79,7 @@ class LogisticRegression():
         else:
             raise ValueError("Cannot handle data with a shape of 3 dimensions or more")
         X_B = np.concatenate((X_data, np.ones((len(X_data), 1))), axis=1)
-        Y_data = np.ones((len(X_data), 1))
+        Y_data = np.ones((len(X_data), 1), dtype=int)
         Y_data[sigmoid(X_B.dot(self.Weights)) < 1 / 2] = -1
         return Y_data
